Initial commit of the Weather Indication Measurement Project - WIMP
This uses a SparkFun MicroMod Weather Carrier Board 16794 along with an
ESP32 based MM board to measure, well, various ambient weather
conditions with an eye to doing so outside.  Additional hardware include
a PMS5003 AQI sensor.

This is an initial release, it's effectively what I have running in
"production" right now.  It has warts, straight up, but it's time to let
others some and look and play

Signed-off-by: John 'Warthog9' Hawley <warthog9@eaglescrag.net>