Swtich back to umqtt.simple - lots of error checking
So I was running into an issue where the station would wander off
and just stop being responsive.  No clear indication of 'why'.
Replicating it on a spare esp32 it looks like umqtt.robust basically
try/excepts the errors, doesn't pass them up, and can easily get stuck
looping indefinitely - obviously this is not ideal.  Switching back to
umqtt.simple resolves this as it traps nothing, and allows us to trap in
main.py.

Adding a lot of try/except blocks to try and make this slightly more
robust, downside is this chews up a LOT more memory - going to need to
mpy-cross the libraries to make this work going forward, and a bunch of
this functionality should likely leave main.py as a result.

Also removing the current mpy in the tree as I had forgotten that that
needs to be re-done per-version used, so this one isn't terribly useful
to keep in the tree

Signed-off-by: John 'Warthog9' Hawley <warthog9@eaglescrag.net>